Refactor student ID caching mechanism
- Moved student ID caching to a separate function in function/cache.py
- Removed local student ID caching implementation from main.py
- Simplified student ID management
