+++ main.py
@@ -2,7 +2,7 @@ from Administrator.menu import administrator_user_page
 from function.query import *
 from function.account_management import register_account
 from Student.menu import student_user_page
-cache_student_id = None
+from function.cache import set_student_id
 
 def get_user_account_type(username):
     accounts = fetch_data("data/user_data.txt")
@@ -14,14 +14,6 @@ def get_user_account_type(username):
             break
 
     return accountType
-
-def get_student_id():
-    global cache_student_id
-    return cache_student_id
-
-def set_student_id(student_id):
-    global cache_student_id
-    cache_student_id = student_id
 
 def check_account_credentials(username, password):
     accounts = fetch_data("data/user_data.txt")
